requirements.txt
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
anthropic==0.18.1
biopython==1.82
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1